Restructured packages to better represent namespaces, began moving elements out of one large file, designed separation of literal RSS data (values may be absent) and effective RSS data (absent values have defaults instead)
