+++ src/main/kotlin/com/github/magneticflux/rss/Persisters.kt
@@ -1,17 +1,37 @@
 package com.github.magneticflux.rss
 
-import com.github.magneticflux.rss.itunes.ITunesBlock
-import com.github.magneticflux.rss.itunes.ITunesBlockConverter
-import com.github.magneticflux.rss.itunes.ITunesComplete
-import com.github.magneticflux.rss.itunes.ITunesCompleteConverter
-import com.github.magneticflux.rss.itunes.ITunesDuration
-import com.github.magneticflux.rss.itunes.ITunesDurationConverter
-import com.github.magneticflux.rss.itunes.ITunesImage
-import com.github.magneticflux.rss.itunes.ITunesImageConverter
-import com.github.magneticflux.rss.itunes.ITunesSubCategory
-import com.github.magneticflux.rss.itunes.ITunesSubCategoryConverter
-import com.github.magneticflux.rss.itunes.ITunesTopLevelCategory
-import com.github.magneticflux.rss.itunes.ITunesTopLevelCategoryConverter
+import com.github.magneticflux.rss.namespaces.default.converters.CategoryConverter
+import com.github.magneticflux.rss.namespaces.default.converters.ChannelConverter
+import com.github.magneticflux.rss.namespaces.default.converters.CloudConverter
+import com.github.magneticflux.rss.namespaces.default.converters.EnclosureConverter
+import com.github.magneticflux.rss.namespaces.default.converters.GuidConverter
+import com.github.magneticflux.rss.namespaces.default.converters.ImageConverter
+import com.github.magneticflux.rss.namespaces.default.converters.ItemConverter
+import com.github.magneticflux.rss.namespaces.default.converters.RssFeedConverter
+import com.github.magneticflux.rss.namespaces.default.converters.SourceConverter
+import com.github.magneticflux.rss.namespaces.default.converters.TextInputConverter
+import com.github.magneticflux.rss.namespaces.default.elements.Category
+import com.github.magneticflux.rss.namespaces.default.elements.Channel
+import com.github.magneticflux.rss.namespaces.default.elements.Cloud
+import com.github.magneticflux.rss.namespaces.default.elements.Enclosure
+import com.github.magneticflux.rss.namespaces.default.elements.Guid
+import com.github.magneticflux.rss.namespaces.default.elements.Image
+import com.github.magneticflux.rss.namespaces.default.elements.Item
+import com.github.magneticflux.rss.namespaces.default.elements.RssFeed
+import com.github.magneticflux.rss.namespaces.default.elements.Source
+import com.github.magneticflux.rss.namespaces.default.elements.TextInput
+import com.github.magneticflux.rss.namespaces.itunes.converters.ITunesBlockConverter
+import com.github.magneticflux.rss.namespaces.itunes.converters.ITunesCompleteConverter
+import com.github.magneticflux.rss.namespaces.itunes.converters.ITunesDurationConverter
+import com.github.magneticflux.rss.namespaces.itunes.converters.ITunesImageConverter
+import com.github.magneticflux.rss.namespaces.itunes.converters.ITunesSubCategoryConverter
+import com.github.magneticflux.rss.namespaces.itunes.converters.ITunesTopLevelCategoryConverter
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesBlock
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesComplete
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesDuration
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesImage
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesSubCategory
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesTopLevelCategory
 import org.simpleframework.xml.convert.Registry
 import org.simpleframework.xml.convert.RegistryStrategy
 import org.simpleframework.xml.core.Persister
