+++ src/main/kotlin/com/github/magneticflux/rss/namespaces/itunes/converters/ITunesConverters.kt
@@ -1,10 +1,16 @@
-package com.github.magneticflux.rss.itunes
+package com.github.magneticflux.rss.namespaces.itunes.converters
 
 import com.github.magneticflux.rss.DurationTransform
 import com.github.magneticflux.rss.URLTransform
-import com.github.magneticflux.rss.children
-import com.github.magneticflux.rss.fallbackPersister
-import com.github.magneticflux.rss.fullName
+import com.github.magneticflux.rss.namespaces.default.converters.children
+import com.github.magneticflux.rss.namespaces.default.converters.fallbackPersister
+import com.github.magneticflux.rss.namespaces.default.converters.fullName
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesBlock
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesComplete
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesDuration
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesImage
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesSubCategory
+import com.github.magneticflux.rss.namespaces.itunes.elements.ITunesTopLevelCategory
 import org.simpleframework.xml.convert.Converter
 import org.simpleframework.xml.stream.InputNode
 import org.simpleframework.xml.stream.OutputNode
